requirements.txt
opencv-python
baidu-aip